+++ plugins/dom.js
@@ -4,9 +4,9 @@ export default function (callback) {
   const parser = new DOMParser();
 
   return (site) => {
-    site.afterRender([".html"], transform);
+    site.process([".html"], processor);
 
-    async function transform(page) {
+    async function processor(page) {
       const document = parser.parseFromString(page.content, "text/html");
       callback(document, page);
       page.content = document.documentElement.outerHTML;